Cleanup zep changes
- Reverts some of the zep changes made in previous commits
- Uses event instead of calling directly into syntax from buffer
- Fix SetBufferSyntax to use name of provider instead of file extension
-
