+++ contrib/zep/include/zep/syntax.h
@@ -59,7 +59,6 @@ public:
     }
 
     virtual SyntaxResult GetSyntaxAt(const GlyphIterator& index) const = 0;
-    virtual void QueueUpdateSyntax(GlyphIterator startLocation, GlyphIterator endLocation) = 0;
 
     const uint32_t ToBackgroundColor(const SyntaxResult& res) const
     {
